requirements.txt
openai-whisper
ollama
python-dotenv
requests
streamlit
markdown
xhtml2pdf
groq
